style fixes and cleanup
